[API] Universal kernel MMIO memory mapper
This is a unification of the kernel MMIO memory mapper over the previous
slightly divergent platform-specific implementations. It allows to map
physical memory with page granularity and to also unmap previously
mapped physical memory (if the memory management subsystem is already
fully initialized). It is also thread-safe which is required to map and
unmap MMIO memory at run time.

If a specific platform needs a custom implementation, the universal code
can be partially or fully overridden.

There is a temporary compatibility wrapper for the old MMIO memory
remapper to ease the transition with respect to argument and return
types.

Change-Id: Id189ed2ebde75911fb2ee8538cf285f9c84dc9d0