spin_lock/arm32: Simplify and unify Spin_lock::lock_arch()
The Spin_lock::lock_arch() implementation and the assembler copy of it
in tramp-mp.S differed significantly on arm32. For example, the copy was
missing a dmb at the end, so might even have been incorrect. Avoiding
the code copy altogether by calling lock_arch() from tramp-mp.S is
unfortunately not an option, since in tramp-mp.S we have no stack.

Furthermore, the non-exclusive load at the beginning of
Spin_lock::lock_arch() appears unnecessary, so remove it.

Change-Id: I566a9dead77566a3f07d73227ded3a36089f2ecc