Explicit reply capabilities
Add the reply capability space to the "Space". Just like the regular
object capabilities, we're holding the reply capabilities in a
two-level table.

Because reply capabilities are now addressable by other threads, care
must be taken when using a reply capability slot. Basically, all
operations on reply cap slots need to be atomic.

Change-Id: I2a8e4348dc8416247c86dbc25eda0c44a4a3ff08
Signed-off-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>
Signed-off-by: Jan Klötzke <jan.kloetzke@kernkonzept.com>
Co-authored-by: Jan Klötzke <jan.kloetzke@kernkonzept.com>