ia32/arm: Timer::system_clock() reads timer with CONFIG_SYNC_CLOCK=y
On ARM, if CONFIG_ARM_SYNC_CLOCK is enabled, Timer::system_clock() now
always reads the generic timer rather than KIP clock value. Prior to
this change, it used the generic timer on the boot CPU and the KIP clock
value on AP CPUs.

On x86, if CONFIG_SYNC_TSC is enabled, Timer::system_clock() now always
reads the time stamp counter directly rather than the KIP clock value.
Prior to this change, it used the time stamp counter on the boot CPU and
the KIP clock value on AP CPUs.

Userland does so as well in these configurations.

Change-Id: Ic3e55003487a7bb3f308321caef7fe078809085d