perf_cnt: explicit config option and enable fixed-function counters
On processors with support for architectural performance monitoring,
enable all corresponding counters, usually 3 or 4, for all read access
from all privilege levels. This enables to use

  RDPMC(ecx=(1<< 30) | x)

for x={0,1,2} on such processors.

CONFIG_PERF_CNT depends on JDB. So far support for performance counters
was automatically enabled if JDB is enabled. Make this option explicit
because with enabling fixed-function counters for all privilege levels,
it might not be desired that userland has a common time base.

Change-Id: Iab2e7990aab8608d56659cecb80c7bbb54ad218f