utest_fw: add Utest::timer_interrupt_indicator()
This function returns a value which changes when the timer interrupt was
triggered. If the value returned by two invocations did not change, no
timer interrupt was triggered between the two invocations.

Usually this function just reads the KIP clock but accessing the KIP
clock directly doesn't work with certain configurations.

Change-Id: Id3f4db22f5716f75d325baf69295a9dabf4a103c