Timer_tick: don't warn about Vlog IRQ on UART with serial hack
The code introduced in I964eb1305131ebe2cb97dfe066747f41daf71654 is
supposed to warn if userland attaches to the Vlog IRQ but the kernel
does not call Vkey::check_(). Improve the code to handle the case if
the "serial hack" is used: In that case, Vkey::check_() is called from
Timer_tick::handle_timer() and the warning is wrong. The handling in
Timer_tick::Timer_tick() was wrong because this constructor is only
called in certain cases.

Change-Id: If7e1ded21dd00e7e37efa2d934929750e502e573