irq: do not loose edge triggered interrupts if unbound
If an Irq_sender is bound to an ICU with an edge triggered interrupt,
we previously were prone to loosing interrupts. An interrupt that hits
while the Irq_sender was not bound to a target thread, was dropped.
This is a problem on edge triggered interrupts because they are lost.

Now, if an edge triggered interrupt hits without a target thread, the
internal send state is marked as "pending". Once a thread is bound, the
IPC is delivered. The pending state is also set if the Irq_sender is
detached with a pending IPC send.

Change-Id: Ic364f34afd08e3397cc61c0015c46415a5cd8b8e