src/Makefile: Remove scripts too on cleanall
We need a way to cleanup mconf and friends without also removing the
configuration files. When the build environment changes, e.g. system
updates or when moving a build directory to another host or another
build environment, and the original libraries mconf was built with do
not exist anymore it will not be rebuilt automatically, but throws
errors and exits.

Change-Id: I50031f9ca0f1e3b71c9c2fa5e557f77954d8ca76