sandbox: Do not pull in Makeconf in pre-sandbox make
When calling make without being in the sandbox already the compiler
might not yet be available. However, Makeconf assumes this to be the
case as it tries to determine the type of the compiler and other flag
support.

Duplicate the VERBOSE logic to support verbose printing, while not
pulling in Makeconf.

Change-Id: I3e4cd1fb47012313c82fe3c1e5dd1aba35b16129