ARM64: disable EL1 access to FEAT_LOR system registers
Disable access to the corresponding _EL1 registers for EL1 for the
FEAT_LOR feature.

Also use Cpu::hcr(value) to set the HCR value because this function
correctly considers the 64-bit size of HCR starting with ARMv8.

There is no TLOR bit in HCR2 so AArch32 is not affected.

See ticket #CD-1315.

Change-Id: I25c1649376116811414cf7cc0d394cbbc7dd45ac