task: fix gcc-14 warnings about uninitialized quota in delete operator
This change fix warnings triggered by gcc-14 because the compiler
assumes that the _quota field is not initialized (anymore) in the delete
operator.

The compiler may assume that the object is not valid anymore when the
delete operator is called, hence, the compiler warns if object members
are accessed from there. This change prevents the warning by telling the
compiler that the to-be-deleted object has some defined content.

Addendum to If2f663de6a3141a8958204e51aadff2319822b22

See also ticket #CD-1323.

Change-Id: I1ddb49c1cc3a96c48dc0ae94f6aec855c2cd5cb9