tb_entry: absolute timeout for 32-bit systems and cleanup
The space for a trace buffer record is very limited on 32-bit platforms
(64 bytes versus 128 bytes on 64-bit platforms) so there is normally no
space available to store a 64-bit absolute timeout in Tb_entry_ipc.
However, if there is no send phase we may re-use the unused words[0,1]
to store that information.

Absolute timeouts are more used for the receive case anyway (timer
thread sleeping until the next timer period).

While at it, prepare for absolute send timeouts in case we ever want
to store them and cleanup the format_timeout() function.

Change-Id: I12e9be721a01aebb1ea6a476c212bde9f447e35a