arm/smmuv3: Avoid blocking CPU uninterruptibly while configuring SMMU
The SMMU uses a command queue as its configuration interface. To ensure
that for example the configuration of a DMA domain association or a TLB
flush is complete, we have to wait until the SMMU has completed the
execution of the configuration commands that we submitted to the command
queue. Sometimes such a configuration consists of several sub-steps,
where one must be completed before the next can be started. Waiting for
completion is implemented using polling, i.e. the executing thread
repeatedly polls the command queue until the commands are completed.

Prior to this change, to coordinate concurrent binding and unbinding of
DMA domains, we used a per-SMMU spinlock which is held for the entire
operation. However, this approach does not work well with the command
queue interface mentioned above, as it means that while waiting for our
submitted commands to complete, we block the CPU uninterruptibly. The
same applies to the unlikely case that the command queue becomes full
and we have to wait for command slots to become free.

Therefore, we replace the per-SMMU lock with a stream table entry state
machine and multiple more fine-grained locks:
  - Lock per DMA domain (introduced in a preceding change)
  - Lock for stream table allocation (introduced in a preceding change)
  - Lock for stream table invalidation

A stream table entry (STE) can be in one of three states: Invalid,
Exclusive, or Valid. To bind or unbind a DMA domain to/from an STE, we
transition the STE into Exclusive state. Then we have exclusive access
to the STE and no one else can modify the STE until we are done with
configuring (transition to Valid state) or deconfiguring (transition to
Invalid state) the STE.

Concurrent operations on the same stream table entry (stream ID) are not
supported. If user-space does that, the kernel might respond with
L4_err::EBusy. More precisely, it does so when it encounters an entry in
Exclusive state.

See ticket #CD-944.

Change-Id: I3500950d97c21837662d7ee76aec57cdbf5724b3