jdb: add call to print kernel memory statistics
The 'km' JDB command prints the kernel buddy allocator memory
statistics. Add a dedicated JDB capability call to do the same.

Especially on no-MMU systems it is not always an option to include JDB
in the build. Also, JDB itself requires memory which influences the
measurement. Having it as a separate call minimizes the impact.

Change-Id: I99819d202f1398db31d2b3a98c200e1ec6beb52c