arm/smmuv3: Introduce wrapper for stream table entry pointers
The wrapper ensures that all accesses to stream table entries in the
SMMUs stream table are done with volatile semantics. In addition it
provides functions to make certain atomic changes.

This change is in preparation for the upcoming removal of the big lock
that was held over the entire bind/unbind/remove operation.

See ticket #CD-944.

Change-Id: I0a09806ca4ac14bc4f0fbab0b40107c164c850e4