utest: make unit test payload preemptible
It might happen that the unit test functions need to sleep, e.g. due to
a blocking DRQ. Because the workload is executed in the idle thread,
sleeping is not allowed. To fix this, create a dedicated thread and
execute the unit test workload there.

Executing the kunit tests in a separate thread is closer to reality
because in normal operation, the kernel code also runs in separate
threads and not in the idle thread.

By naming the entry point for tests init_unittest_exclusive() instead of
init_unittest(), no dedicated thread is created.

Change-Id: Iedc43197ca54b51400cf96edb25086a0a11f5699