utest_fw: keep correct stack alignment in start_thread
Before performing Context::prepare_switch_to() to set the instruction
pointer of the new kernel thread, do Thread::reset_kernel_sp() to reset
the stack pointer of the target thread. On some platforms, for example
ARM32, the stack pointer was changed when the Thread constructor called
Context::prepare_switch_to().

This ensures that the stack pointer for the new thread is correctly
aligned.

Change-Id: Ia4767e0276bb8fd2e81e1443ebaf7354926c15b3