Raise Kobject lists abstraction, remove triple and double pointers
Raise the abstraction level of Kobject lists (currently used by the reap
lists) by introducting a helper class that encapsulates and documents
the logical operations with the lists.

This also enables to get rid of the double and especially of the triple
pointer arguments that made the code quite hard to understand before.

Acknowledgement: The code is inspired by a prototype implemented by
Nataliia Korop <n.corop08@gmail.com> (used with permission), but
modified and extended.

Change-Id: Idade640454c2b0b848facdfcf8b0e6436ee62a5d