preprocess: allow c++ function attributes
Function attributes that belong to the function and not to the type
must go first in the function declaration (after a potential template
definition) according to the C++ standard.

Adapt preprocess to capture attributes from the input source and make
sure to make the attributes the first parts of the declaration right
after potential template statements. Also add test cases for
the corresponding source cases.

Change-Id: I0ce2ae50e8a61333f8dd4dcff7c2967fe51671c5