arm: keep KIP read-only on MPU systems
Since I7e52b87be22fc235027c4d7dba92716eedaefdd4 the KIP clock value is
not updated with CONFIG_SYNC_CLOCK=y. Force the feature to be enabled
on all R-profile platforms. This saves us some cycles because we do not
have to make the KIP writable on kernel entry and read-only on kernel
exit.

Change-Id: Ide8c31fd7c04980a6882911d4040d4e0dda7062d