lock_guard: acquire lock only in constructor
This prevents accidental double-locking attempts and accidental implicit
release() when calling lock() with two different locks.

Additionally, create the sub-class Switch_lock_guard that provides the
member is_valid() for checking if the lock state is valid. In contrast
to Lock_guard, the sub-class enforces initialization of the state in the
constructor, so it is always safe to call is_valid().

Change-Id: I75271426d5100e2f471e3ca8ce08a9ffd9812ba3