Remove lock module
The lock module provides the rather generic Lock typedef indirection
for Helping_lock. This indirection does not offer any benefit, but is
actually rather detrimental, as it obscures what kind of lock is behind
the generic Lock.

The Helping/Switch_lock is not a conventional lock, it has some
pitfalls, so its use should be clearly visible.

Change-Id: Ib95c94228fc45a8753e825e07fa46c4ce3234bc2