arm-fvp_base_r: Implement shutdown & reboot
The ARM FVP has a system register device which can be used for shutdown
& reboot among other things. This change implements the necessary steps
to trigger a shutdown or reboot on the ARM FVP BaseR (without MMU)
platform.

The device is implemented within the drivers section of the source code
so it can be reused for the ARM FVP Base platform (with MMU).

Change-Id: I375cf4311c9f69630dbf4a7b3f8d5c0b503c45c2