utest-fw: Add shutdown as action to do once the kunit test is done
In some situations we want the unit test framework to initiate a
shutdown, instead of a reboot, after the current test has finished
testing. To facilitate this we replace the Boolean field restart in the
External_info structure with an enum field that can either be None,
Reboot or Shutdown.

That enum field can be set directly from the command line, i.e. '0' for
None, '1' for Reboot, '2' for Shutdown. For backwards compatibility
reasons we've chosen '1' as the value for Reboot. The enum can be easily
extended with other actions. Specifying any invalid value simply behaves
like None.

Furthermore call Platform_control::system_reboot() instead of
platform_reset() because the former by default calls the latter.

Change-Id: I58ff4c0cdc74a299c0f0f28348a6062f30a3e6d3