[API] ia32: Encode userland IOMMU selection in SRC ID
To assign a device to the IOMMU on the device's bus, it is necessary to
parse the DMAR table entries and in case of substructures the referenced
PCI bridge. PCI bridge parsing is not the duty of the kernel and must be done in userland.

Thus, this requires an interface to communicate the information of
the IOMMU to use in the IOMMU bind/unbind API.
We extend the source ID parameter to encode the index into the ACPI DMAR
table. Since kernel and userland process the same table, the index is
sufficient to identify the IOMMU.

Furthermore, this removes the bus matching extension, since it
contradicts the removal of device to IOMMU assignment in the kernel.

fixes ticket #L4RE-412

Change-Id: Iec87d9ed33d666497ae5017beef685064ba36216