mpu: improve user space region initialization
Context switches between tasks need to reprogram the MPU. This only
involves user space regions but a couple of regions in the MPU are used
by the kernel too. These regions are managed in the kernel task and do
not change after the initial mappings were established by the kernel.

Instead of keeping these things completely separated, copy the region
data from the kernel task to new user space tasks every time. All
reserved and used regions of the kernel task are made reserved in the
new user space task.

By copying the region data, the context switch code does not need to
care when writing kernel regions. This enables some optimizations in
the following commits.

Change-Id: I06d5cc673a5298a7f15c8f06387c78f4077ff9b3