arm64: manually unroll mpu update loop
So far the compiler was not really able to infer some loop constants.
Instead, gcc always computed the offset to each region struct. By
manually unrolling the loop the generated code looks much better. It
now uses offsets to a base register as expected.

This enables a further optimization to skip the update of the three
well known regions (kernel, KIP, heap). Their value is known to be
constant.

Change-Id: Icf5d6f964a46bd812fb3f2c1ee8402e0a12e1ce1