lib/gcov - Implement output via memory buffer
Up until now we only supported coverage output via stdio, which for most
cases works. In case of the ARM FVP simulator, pushing out coverage data
via the UART is much too slow. So in order to export the coverage data
from this simulator in a timely manner, we need another way to transfer
the data to the host system.

The ARM FVP simulator has a feature to dump a section of the memory into
a file after shutdown. To utilize this for coverage we add a feature to
libgcov where it will write the coverage data into a memory buffer
instead. The memory buffer is designated by a physical start address and
a buffer length, both of which are freely configurable via Kconfig.

To avoid implementing any further decoding on the host side, we simply
write the same base64 encoded data stream to the memory buffer that
would've been written to stdio.

Change-Id: If23f956e49e43ee4db94b3a8c7447366a310ebb4