mpu: drop reserved test when writing regions
Now that user space tasks have a copy of all kernel regions, there is
no need to skip reserved regions anymore. This improves the generated
code quite a bit because there was a test necessary for each region.
These were many conditional branches in the hot context switch path
that are now gone.

Change-Id: I4eaff2fc55ceaf883167dbd9277f00956921526c