jdb_tcb: improvements for dumping kernel stack and TCB
The following applies only for creating a TCB/stack dump when doing
reports ('D') or when explicitly performing such a dump ('tcbdump'):

Do not only skip stack content containing zeros but also skip most of
the content below the stack pointer even if it does not contain zeros.
While the TCB memory is initialized to zero, the stack is not and may
contain left-overs from the system boot process. This non-initialized
content enlarges the report but does not provide useful information so
skip it. Nevertheless 5 lines starting from the kernel SP downwards are
always printed.

While at it, insert a few CRs when printing the report to improve the
readability.

See also ticket #CD-1298.

Change-Id: Ibfeea522cdc4cf84546b8030087278af7de77f37