Intel IOMMU: More resilient queued invalidation setup
Disable queued invalidation before enabling it for a more resilient
setup. If the queued invalidation has been already enabled before (i.e.
by the firmware or by the boot loader), this is the only way to reset
the head pointer to 0 and thus avoiding undefined behavior of the
hardware.

Specifically, this fixes booting via the kexec mechanism.

Change-Id: I3cd190a24c895ab65e0cf62406839d6d8c34a63d