arm: sbsa: improve redistributor probing
There are two ways to find the GIC redistributor of a CPU on Arm ACPI
platforms: via a "GIC Redistributor (GICR) Structure" or from the "GIC
CPU Interface (GICC) Structure" directly.

So far, Fiasco only supported a single "GIC Redistributor (GICR)
Structure". In fact, there can be more than one as per specification.

More importantly, the GICR structure is not usable on all platforms.
Only if all redistributors are in the always-on power domain, such a
structure can be used. On all other platforms, the GICC structure hold
the pointer to the redistributor. So if no GICR structure is present,
we use the GICC redistributor field.

Change-Id: I6d5e63f626666397829e6e3fde05fe3da80a6084