kobject_mapdb: ensure consistent lock ordering
If an object capability is mapped and the destination task already has
an object capability at the destination slot, both underlying kernel
objects are locked. This might happen concurrently for the same
objects, but with source and destination swapped. To guarantee a
consistent lock ordering, use the object address as criteria which
object is locked first.

Fixes a potential deadlock if multiple threads map the same underlying
kernel objects but with source/destination swapped.

See ticket #CD-1378.

Change-Id: I5d5772551c93cce64f6689490c37a936f967c2cb