Add template for noreturn checks
When functions do not return the code afterwards is unreachable.
Sometimes we used to have some defensive programming code to take care
if for whatever reason the code did return. This causes issues with dead
code warnings and is also hard to maintain since the code cannot be
tested.

The issue is that, should a function change from noreturn to actually
returning without adapting the call-site this can lead to undefined
behavior.

To help with this we introduce a check_noreturn template to cxx that
acts as wrapper, and itself is a noreturn function. Should the wrapped
function ever change its attribute to return something then the wrapper
will cause a warning message at the call-site allowing developers to
adapt the code.

Make use of this in places where we previously relied on the noreturn
behavior.

This is a pre-requisite to enable unreachable-code warnings.

Change-Id: I04bf82de18229d621347b7333a0cab9c09743637