Update cursor position on output
Fix visual imperfection when the cursor position was not updated on
output (but only on input), keeping the cursor at an awkward position
not corresponding to the last line of output.

To avoid any major performance impact, the cursor position is only
updated once after outputting an entire chunk of characters.

Change-Id: I6ca20f7e715d18330715c900a0f2029427ceaaee