platform_control/ia32: Restore spilled FPU state on resume
After returning from a system suspend Fiasco only initializes the FPU,
but does not restore the FPU state of the current thread, which was
spilled (saved) before entering the suspend.

Before Iaa7a044890678c3c817b0e287f982e2d99b344ef that did not matter,
because do_system_suspend() was always executed by the kernel idle
thread of the boot CPU, whose FPU state is irrelevant. But now that
do_system_suspend() is executed by a regular thread, that also executes
in user mode, the FPU state of this thread doing the system suspend is
relevant and must be restored.

However, we only have to take action with eager FPU switching, because
there the FPU always belongs to the current thread, whereas with lazy
FPU switching the FPU just has no owner, and is then restored later on
demand.

See ticket #CD-1384.

Change-Id: Iebf2735abed69ef53b6cd7256e321ca4fc2cdd6e