Remove unreachable code
These are all cases where the code is clearly not reachable either
because panic is called, or because code branches away before the
statement. All removed cases should be safe. If there is ever a way
introduced such that the location is reached then a warning (such as
missing return) is issued by the compiler.

Change-Id: Ifda0d1b7b4b633191fba9b4e0004f394dd4cc37e