Test framework for code rules adherence
This is a generic framework that allows to add some tests to ensure that
problematic constructs in code are reported on by relevant tooling.
Currently the main supported type of test is a warnings test where we
check that compilers warn about all constructs that are considered
problematic in the coding rules when they are used.

When enabled the tests code is compiled (not linked) and a script is
generated that checks the compiler output against expected warnings
declared inside the test source code.

Change-Id: Iaa36f971062ddf46eb0123f2bbbfd6d3b71862da