Make debug print conditions constexpr
This change makes all places where debug output is guarded by an if
statement such as if (0) constexpr. Where multiple prints are guarded
introduce an enum to have a speaking name for the condition. In
single-use cases keep constexpr if (0) and add a comment to avoid being
overly talkative in the code.

This is a pre-requisite to enable unreachable-code warnings.

Change-Id: I2a35b1acea1ea59ca29c013d1567e76d071850ac