x86: Remove effectively unused getchar_does_hlt_works_ok feature
In commit 25fd3947da9cc49915336641ad3e132b58766a06 from 2013 the
getchar_does_hlt_works_ok flag got disabled, because it did not
work properly in the event of MSIs and IPIs.

The only remaining way to use the getchar_does_hlt_works_ok feature
is with timer-hpet, which only is available in non-smp configuration.
In addition it is tagged as EXPERIMENTAL.

Therefore, the feature can be considered effectively unused and can
be removed.

Change-Id: I25cee09c7cd6cc981b4d50eb889441d36f565edc