Allow collecting diagnostics to files during build
When COLLECT_DIAGNOSTICS is set to a non-empty value then all
diagnostics reported during build are written to a .diag file in the
build directory, where the stem of the filename is the name of the
compile output file.

This is useful for highly parallel build to avoid mangled diagnostics
output (clang) and for getting diagnostics after a build.

Change-Id: I9db7e36db2bc123f7a1da4a87fb2e40661059c76