Adapt fallthrough attributes and enable warning
Clang does not support automated attribute parsing for eliding the
implicit fallthrough warning. Use the language attribute instead.

For C the [[fallthrough]] attribute is only supported since C23 which is
not yet widely supported. So we stay with the compiler specific
__attribute__((fallthrough)) for now.

This should adapt all missing places and allows us to also enable the
corresponding warning.

Change-Id: I05ac9523772a072beffba4d24ab7b6ececf7b311