irq: Refactor irq_direct_inject conditions
The is_vcpu_irq condition is constexpr in cases where we do not support
direct IRQ injection. However, in cases where we do support it, it is
dependent on the thread flags, so cannot be constexpr.

To avoid dead code warnings in these cases (and make clear that the dead
code is intentional) factor out the condition check if irq_direct_inject
is enabled into an extra constexpr condition. Thus we only need to call
is_vcpu_irq check in cases where it is actually a non constexpr
function.

Change-Id: I7ec062256c44b144963e5ff8df3f3b6ebd2b9452