[API] IPC gate: error when binding with label with a lower bit set
The lower bits of an IPC gate label are used for signaling permissions
to the receiver. Therefore return an error when a user tries binding
with a label where the lower bits are not zero. This avoids that a user
accidentally sets those bits and gets wrong permissions signaled later.

Change-Id: Ie44f92bb563fbf7b8d45d378547ae0c39b9994fa