tickless_idle: Ignore idle thread timeslice timeout
To remain in tickless idle for as long as possible, we ignore the
timeslice timeout of the idle thread when determining the next timeout
for which we have to wakeup from tickless idle.

This is safe to do because we know that no other threads with the same or lower priority than the idle threads exist in the system.

Change-Id: Ie2870b1d5b164cbb907f2aff898356670c94f833