arm: make iommu array dynamically sized
On the Arm SBSA platform the number of IOMMUs is not known at compile
time. So instead of allocating a fixed array, let the platform code
allocate an array with the actual number of IOMMUs.

The CONFIG_ARM_IOMMU_NUM, resp. Iommu::Num_iommus constants thus become
the upper limit of supported IOMMUs. For most platforms that makes no
difference because CONFIG_ARM_IOMMU_NUM and the allocated array size
are hard coded.

Change-Id: I3da065774e403b71507f5b5a56df0596079470b4