UART and DRIVERS may depend on LIBK
Code in DRIVERS might depend on code in LIBK, for example for on code in
gcc_lib. Reflect this dependency in the order of libraries passed to the
linker when creating the kernel image.

This fixes a build error on arm. Change the order on all similar rules
to keep the link order consistent over the platforms.

Change-Id: I436826a567ce44bdb2841378d9f3c832351ec680