convert a couple of bool enum constants to static constexpr bool
Address the following Clang warnings (seen with Clang-20):

  ...src/kern/riscv/mem_space-riscv.cpp:520:28: warning:
  use of logical '||' with constant operand [-Wconstant-logical-operand]
  520 |   if (Mem_unit::Have_asids || Mem_unit::Have_vmids)
      |                            ^  ~~~~~~~~~~~~~~~~~~~~
  ...src/kern/riscv/mem_space-riscv.cpp:520:28: note:
  use '|' for a bitwise operation
  520 |   if (Mem_unit::Have_asids || Mem_unit::Have_vmids)
      |                            ^~
      |                            |

Using '|' instead of '||' is actually the same but makes the code harder
to understand. It does no matter if we use

  enum : bool { Have_xyz = true };

or just an untyped enum. Also using

  if constexpr (Mem_unit::Have_asids || Mem_unit::Have_vmids)
    ...

doesn't make a difference. But using 'static constexpr' to convince the
compiler is an acceptable alternative. While the compiler only warns for
RISC-V, also change a few more constants like this.

Change-Id: I7d6abf015dd5b2c582d3924b15faaa4e66e595b5