mips: adjust PC on entry rather than on exit
For syscalls, the PC needs to be adjusted to point after the
instruction. This needs to be done on exception entry rather than the
exit to have a consistent view onto the program counter while a thread
is in the kernel. In particular, if the program counter is modified
(e.g. via ex_regs() or in the exception IPC handler) while the thread
is blocked in a syscall, the execution must resume on the newly set
instruction. Previously the new PC was adjusted by 4, leading to
unpredictable behaviour because that would only happen in the syscall
path and not on any other kernel exit path.

Requires-Change-Id: I742b47f57041588c19aba990c41a75883ad89117
Change-Id: I19754d1c9eb79cc09ef60a5fcb1b2eea11ce6e9d