templates: Remove SCHED_WFQ=y from amd64-mp
The WFQ scheduler is experimental and by default userspace expects the
fixed priority scheduler. Selecting WFQ causes moe to crash in a way
that is really confusing. When moe tries to set the scheduling
parameters for the newly started thread (probably l4re_itas), the kernel
rejects them as invalid. This is because moe provides fixed-priority
scheduler parameters and not WFQ scheduler parameters.

Change-Id: I9274160cd00d41d2d9570a3230e601629bc501ee