thread-ipc: Avoid race between setting and clearing Thread_ipc_transfer
When a sender tries to abort an IPC on a remote CPU that is already in
progress, i.e. the receiver is currently transferring the IPC message
through Thread::ipc_send_msg(), the sender receives the result
Abt_ipc_in_progress from Receiver::abort_send().
Thereupon it sets the Thread_ipc_transfer flag and waits until the
receiver clears it, when it has completed the transfer.

Unfortunately there is race condition in that, so it can happen that the
receiver clears the Thread_ipc_transfer flag before the sender sets it,
for example, because other threads on the sender CPU prevent the sender
from running for some time. The consequence of the race is that the
sender loops endlessly in "while (state() & Thread_ipc_transfer)".

To prevent this race condition, the Thread_ipc_transfer flag must be set
already in Receiver::abort_send() via xcpu_state_change(), i.e. in the
context of the receiver, thereby eliminating the possibility of a race.

See ticket #CD-1454.

Change-Id: Iae20f8e847d748d317877c262fe3de11183848b6