Reduce initial build delay due to unnecessary recipe execution
Currently, in each make run the recipes for all generated source files
are executed. This causes a noticeable delay (about a second) before the
build actually begins. Some background to understand what is going on:

The make recipe that generates source files with preprocess "only
updates the generated files if they actually change; that is, if you
change something in the implementation section that does not influence
the header files, they will not be updated, avoiding recompilation of
files that depend on them".

The way this is implemented is by having a ready stamp target, which has
the source files as prerequisites, and itself is the prerequisite for
the generated source files. The recipe of that ready stamp is
responsible for invoking preprocess to create the generated source
files. It uses the move-if-change script to only update a generated
source file if it changed.

That can lead to a situation, where the ready stamp is newer than one
(or multiple) of the generated source files. In such a case, the recipe
of the generate source file is executed, since make sees that its
prerequisite (the ready stamp) is newer, so thinks it is outdated and
needs to be recreated. For this case, the recipe contains a check to see
if the generated file already exists, and if it does, the recipe does
nothing.

Unfortunately, this is not a rare exception, but affects all generated
source files in each make run, because the make recipe that generates
the source files with preprocess, creates the ready stamp only after
generating the source files, i.e. it is always newer than them.

To avoid that overhead of starting all the shell processes, the recipe
now creates the ready stamp before generating the source files. Even if
subsequently preprocess encounters an error while generating the source
files, it should be fine, because preprocess does not directly create
the target files. Only after successful generation the make recipe moves
the generated source files to their destination.

Change-Id: I382974ad9448f600a55db58df4c29310de4f300d